orginized seed routed into specific controls to the models and updated server.js routes
